added pages for user created houses and bookings. refactored some files
