+++ frontend/src/Pages/Dashboard/Dashboard.jsx
@@ -3,16 +3,22 @@ import { authRequest } from "../../utils/requests/auth";
 import { useNavigate } from "react-router-dom";
 import { logoutRequest } from "../../utils/requests/logout";
 import { UserContext } from "../../Context/UserContext";
-import { fetchHousesByOwnerIdRequest } from "../../utils/requests/fetchHousesByOwnerId";
-import { fetchBookingsByBookerRequest } from "../../utils/requests/fetchBookingsByBooker";
-import { Link } from "react-router-dom";
+import Userhouse from "./components/Userhouse";
+import UserBookings from "./components/UserBookings";
+import AddHouse from "./components/AddHouse";
+
+const dashboard_pages = {
+    dashboard: 0,
+    house: 1,
+    bookings: 2,
+    addHouse: 3
+}
 
 export default function Dashboard() {
     const navigate = useNavigate();
     const { user, setUser } = useContext(UserContext);
     const [loading, setLoading] = useState(true);
-    const [userCreatedHouses, setUserCreatedHouses] = useState([]);
-    const [userBookings, setUserBookings] = useState([]);
+    const [dashboardPages, setDashboardPages] = useState(dashboard_pages.dashboard);
 
     useEffect(() => {
         const auth = async () => {
@@ -25,16 +31,7 @@ export default function Dashboard() {
                     email: userData.email,
                 });
 
-                // fetch user created houses if any
-                const houses = await fetchHousesByOwnerIdRequest(userData.id);
-                setUserCreatedHouses(houses);
-
-                // fetch all bookings made by this user
-                const bookings = await fetchBookingsByBookerRequest(userData.id);
-                setUserBookings(bookings);
-                console.log(houses);
                 setLoading(false);
-                console.log(userData.id);
             } catch (err) {
                 console.log(err);
                 navigate("/login");
@@ -60,22 +57,29 @@ export default function Dashboard() {
         return <div>Loading...</div>
     }
 
+    switch (dashboardPages) {
+        case dashboard_pages.house:
+            return <Userhouse setDashboardPages={setDashboardPages} />
+
+        case dashboard_pages.bookings:
+            return <UserBookings setDashboardPages={setDashboardPages} />
+
+        case dashboard_pages.addHouse:
+            return <AddHouse setDashboardPages={setDashboardPages} />
+
+        default: break;
+    }
+
     return (<div className="bg-gray-100 min-h-screen flex flex-col justify-center items-center">
         <h1 className="text-4xl font-bold text-gray-800 mb-6 text-center">Dashboard</h1>
         <div className="w-full max-w-md  p-6 text-center">
-            <Link to="/userhouse" path>
-                <button className="bg-red-500 hover:bg-red-700 text-white font-bold py-2 px-4 rounded-full mt-4 ml-4">
-                    My Houses
-                </button>
+            <button className="bg-red-500 hover:bg-red-700 text-white font-bold py-2 px-4 rounded-full mt-4 ml-4" onClick={() => setDashboardPages(dashboard_pages.house)}> My Houses</button>
 
-            </Link>
-            <hr className="my-4" />
-            {/* Add your content here */}
         </div>
         <div className="flex justify-center">
             <button
                 className="bg-red-500 hover:bg-red-700  text-white font-bold py-2 px-4 rounded-full mt-4 ml-4"
-                onClick={() => navigate("/add-house")}
+                onClick={() => setDashboardPages(dashboard_pages.addHouse)}
             >
                 Add a house
             </button>
@@ -87,7 +91,7 @@ export default function Dashboard() {
             </button>
             <button
                 className="bg-red-500 hover:bg-red-700  text-white font-bold py-2 px-4 rounded-full mt-4 ml-4"
-
+                onClick={() => setDashboardPages(dashboard_pages.bookings)}
             >
                 My Bookings
             </button>
